feat: Lv.2 - 리스트에 MenuItem 삽입

+++ src/main/java/com/example/level2/MenuItem.java
@@ -2,10 +2,10 @@ package com.example.level2;
 
 public class MenuItem {
     private String name;
-    private int price;
+    private double price;
     private String description;
 
-    public MenuItem(String name, int price, String description) {
+    public MenuItem(String name, double price, String description) {
         this.name = name;
         this.price = price;
         this.description = description;
